backend/requirements.txt
fastapi
uvicorn
langchain
google-generativeai
faiss-cpu
PyPDF2
tiktoken
streamlit
python-multipart
